# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: OmniRepo API
  version: "1.0.0"
  description: |
    REST API for OmniRepo — self-hosted artifact repository server.
    Serves OCI/Docker, RPM, APT/Debian, PyPI, Helm, RAW, S3, and Git
    from a single container on one HTTP/HTTPS port.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - bearerAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ApiErrorClass:
      type: string
      enum: [validation, permission, transient, operator_action_required]
      description: Stable client-facing error classification — drives UI icon/copy/CTA.
    ApiErrorEnvelope:
      type: object
      required: [code, message, class]
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*\\.[a-z][a-z0-9_]*$"
          maxLength: 80
          description: Stable machine identifier (e.g. "repo.not_found")
        message:
          type: string
          maxLength: 280
          description: Human-friendly single-sentence message (no internal paths)
        hint:
          type: string
          maxLength: 280
          description: Optional single-sentence remediation guidance
        class:
          $ref: '#/components/schemas/ApiErrorClass'
        incident_id:
          type: string
          maxLength: 64
          description: Server-generated correlation ID (UUID v7) — matches slog request_id and the X-Incident-Id response header
        details:
          type: object
          additionalProperties: true
          properties:
            field:
              type: string
              description: Dot-path of the offending form field (validation class)
            fields:
              type: object
              additionalProperties:
                type: string
              description: Map of field-path to error-code for multi-field validation failures
            retry_after_ms:
              type: integer
              minimum: 0
              description: Suggested retry delay in ms (transient class)
            operator_route:
              type: string
              description: UI path (e.g. "/admin/trivy") for operator_action_required deep-link
            operator_label:
              type: string
              description: CTA label for the operator deep-link (e.g. "Go to Admin → Trivy")
...
    GitDiff:
      type: object
      properties:
        sha:
          type: string
        message:
          type: string
        stats:
          type: object
          properties:
            additions:
              type: integer
            deletions:
              type: integer
            files_changed:
              type: integer
        files:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              status:
                type: string
              patch:
                type: string
    GitBlame:
      type: object
      properties:
        path:
          type: string
        lines:
          type: array
          items:
            type: object
            properties:
              line_number:
                type: integer
              sha:
                type: string
              author:
                type: string
              date:
                type: string
                format: date-time
              content:
                type: string
    GitRef:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [branch, tag]
        sha:
          type: string
...
  responses:
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorEnvelope' }
...
paths:
...
  /projects/{name}/repos/git/{repo}/commit/{sha}:
    get:
      operationId: gitCommitDetail
      tags: [git]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: sha
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single commit with diff
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitDiff'
        default:
          $ref: '#/components/responses/ValidationError'
  /projects/{name}/repos/git/{repo}/blame/{ref}/{path}:
    get:
      operationId: gitBlame
      tags: [git]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: ref
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blame info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitBlame'
        default:
          $ref: '#/components/responses/ValidationError'
...
  /projects/{name}/repos/git/{repo}/refs:
    get:
      operationId: gitRefs
      tags: [git]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branches and tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GitRef'
        default:
          $ref: '#/components/responses/ValidationError'

  # ============================== SEARCH ==============================
...